% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anes2016}
\alias{anes2016}
\title{ANES 2016 data}
\format{
A data.frame with 8135 observations and 2 variables:
\describe{
\item{docid}{a unique document ID}
\item{response}{text of open-ended response}
...
}
}
\source{
\url{https://electionstudies.org/data-center/2016-time-series-study/}
}
\usage{
anes2016
}
\description{
A data.frame with 8135 observations and 2 variables, containing
unique responses to the open-ended question in the ANES (2016):
"what are the most important issues facing the country?"
See /data-raw for pre-processing file.
}
\keyword{datasets}
