% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anes2016_glove}
\alias{anes2016_glove}
\title{GloVe subset}
\format{
A matrix with 1441 rows and 300 columns:
\describe{
\item{row}{each row corresponds to a word}
\item{column}{each column corresponds to a dimension in the embedding space}
...
}
}
\source{
\url{https://nlp.stanford.edu/projects/glove/}
}
\usage{
anes2016_glove
}
\description{
A subset of the Stanford GloVe (300 dimensional) embeddings, consisting of
all words found in the ANES 2016 dataset with at least 5 occurrences.
}
\keyword{datasets}
