% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelProp.R
\name{labelProp}
\alias{labelProp}
\title{Compute relevance scores for a collection of nodes based on a set of seed nodes.}
\usage{
labelProp(
  x,
  seeds,
  method = "rw",
  beta = 0.5,
  bootstrap = FALSE,
  num_bootstraps = 50,
  prop_seeds = 0.5,
  as_list = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{x}{(numeric) if \code{method = rw}, then \code{x} is a symmetric matrix of transition probabilities.
if \code{method = nns}, then \code{x} is a matrix of vector representations.}

\item{seeds}{(list) a list of character vectors defining the classes of interest.
If named, then \code{names(seeds)} will be used to define classes,
otherwise classes will be labeled \code{class1}, \code{class2} etc.}

\item{method}{(character) either \code{nns} or \code{rw}. If \code{nns}, then values will be computed
using cosine similarity, if not values are computed using spreading activation.}

\item{beta}{specifies the extent to which the algorithm favors local (similar labels for neighbors)
vs. global (correct labels on seed words) consistency.}

\item{bootstrap}{(logical) if TRUE, use bootstrapping -- sample a proportion (defined by \code{prop_seeds})
of seeds and re-run algorithm. Required to get std. errors.}

\item{num_bootstraps}{(integer) number of bootstraps to use.}

\item{prop_seeds}{(numeric) proportion of seeds to sample when bootstrapping.}

\item{as_list}{(logical) if FALSE all results are combined into a single data.frame
If TRUE, a list of data.frames is returned with one data.frame per class.}

\item{verbose}{(logical) if TRUE show progress bar.}
}
\value{
a \code{data.frame} or list of data.frames (one for each class)
with the following columns:
\describe{
\item{\code{node}}{ (character) rownames of \code{x}.}
\item{\code{class}}{(character) name of class. If none provided, then
classes will be labeled \code{class1}, \code{class2} etc.}
\item{\code{score}}{(numeric) score assigned to node.}
\item{\code{std.error}}{(numeric) std. error of score.
Column is dropped if \code{bootstrap = FALSE}.}
}
}
\description{
See: https://proceedings.neurips.cc/paper/2003/file/87682805257e619d49b8e0dfdc14affa-Paper.pdf
}
\keyword{labelProp}
