% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_transition_matrix.R
\name{build_transition_matrix}
\alias{build_transition_matrix}
\title{Computes a matrix of transition probabilities.}
\usage{
build_transition_matrix(x, metric = "cosine", threads = 4L)
}
\value{
a V x V matrix of transition probabilities, with V = number of rows in x.
}
\description{
Computes a matrix of transition probabilities.
}
\examples{

transition_matrix <- build_transition_matrix(anes2016_glove, threads = 2)

}
\keyword{build_transition_matrix}
